video_project/requirements.txt
Django>=3.2
djangorestframework
google-cloud-aiplatform>=1.36.0
google-generativeai>=0.3.0
google-cloud-core>=2.3.3
google-api-core>=2.11.1
gunicorn==20.1.0
whitenoise==6.4.0
dj-database-url==1.2.0
psycopg2-binary==2.9.5
requests>=2.31.0